increase payload size header to 4 bytes to allow messages >65K
fix problem with ServerConnection::sendMessage() passing pointer to local stack to async_write()

+++ src/libwatcher/dataMarshaller.h
@@ -142,7 +142,7 @@ namespace watcher
             enum 
             { 
                 header_length = 
-                    sizeof(unsigned short) +   /// Payload length 
+                    sizeof(uint32_t) +   /// Payload length 
                     sizeof(unsigned short)     /// Number of messages
             };
